requirements.txt
albt==1.0.1b1
appdirs==1.4.3
boto3==1.4.4
botocore==1.5.53
click==6.7
docutils==0.13.1
futures==3.1.1
jmespath==0.9.2
packaging==16.8
pyparsing==2.2.0
python-dateutil==2.6.0
PyYAML==3.12
s3transfer==0.1.10
six==1.10.0
